backend/requirements.txt
Flask
mysql-connector-python
pyjwt
python-dotenv
bcrypt
bleach
pycryptodome
Markdown
gunicorn